[#5] Move test file into __tests__ folder
